requirements.txt
pyrogram==2.0.106
TgCrypto
python-dotenv
yt-dlp
aiohttp
aiofiles
